Rename components/AppLink.vue to components/_shared/AppLink.vue
