+++ components/StaticInfo.vue
@@ -12,7 +12,7 @@
 <script lang="ts">
 import Vue from 'vue'
 
-import AppLink from '@/components/AppLink.vue'
+import AppLink from '@/components/_shared/AppLink.vue'
 
 export default Vue.extend({
   components: { AppLink },
