+++ components/_shared/SideNavigation/MenuList.vue
@@ -27,7 +27,7 @@
 <script lang="ts">
 import Vue, { PropType } from 'vue'
 
-import AppLink from '@/components/AppLink.vue'
+import AppLink from '@/components/_shared/AppLink.vue'
 import CovidIcon from '@/static/covid.svg'
 import MaskTrashIcon from '@/static/masktrash.svg'
 import ParentIcon from '@/static/parent.svg'
