+++ components/index/CardsReference/Metro/Chart.vue
@@ -90,7 +90,7 @@ import { ChartData, ChartOptions, ChartTooltipCallback } from 'chart.js'
 import Vue from 'vue'
 import { ThisTypedComponentOptionsWithRecordProps } from 'vue/types/options'
 
-import AppLink from '@/components/AppLink.vue'
+import AppLink from '@/components/_shared/AppLink.vue'
 import DataView from '@/components/DataView.vue'
 import DataViewTable, {
   TableHeader,
